+++ src/TodoList.jsx
@@ -1,11 +1,36 @@
-function AddTodoForm () {
+const todoList = [
+    {
+      id: 1,
+      title: 'wake up on time',
+    },
+    {
+      id: 2,
+      title: 'bathroom routine',
+    },
+    {id: 3,
+      title: 'prepare and eat breakfast'
+    },
+  ];
+
+
+function TodoList () {
     return (
-        <form>
-          <label htmlFor="title">Title</label>
-          <input type="text" id="todoTitle" />
-          <button type="submit">Add</button>
-        </form>
+        <div>
+    <ul>
+    {todoList.map( function (list) {
+      return (
+      <li key={list.id}>
+        <span>{list.title}</span>
+        </li>
       );
+})}
+  </ul>
+  </div>
+    );
 };
 
-export default AddTodoForm;
+
+
+
+
+export default TodoList;
